version 0.0.0.6

+++ src/main/java/com/mreversing/handsfreelistening/Utils/myPcmReader.java
@@ -24,7 +24,7 @@ public class myPcmReader {
         } catch (FileNotFoundException e) {
             e.printStackTrace();
         }
-        short[] data = new short[(short) len];
+        short[] data = new short[(int) len];
         for (int i = 0; i < len; i++) {
             data[i] = readShort();
         }
